add walk variants util
